+++ encryption/index.js
@@ -1,13 +1,17 @@
 const decryptCiphertext = require('./decrypt_ciphertext');
 const decryptPayload = require('./decrypt_payload');
+const decryptWithNode = require('./decrypt_with_node');
 const derAsPem = require('./der_as_pem');
+const encryptToNode = require('./encrypt_to_node');
 const encryptToPublicKeys = require('./encrypt_to_public_keys');
 const pemAsDer = require('./pem_as_der');
 
 module.exports = {
   decryptCiphertext,
   decryptPayload,
+  decryptWithNode,
   derAsPem,
+  encryptToNode,
   encryptToPublicKeys,
   pemAsDer,
 };